+++ build.py
@@ -13,7 +13,7 @@ sys.path.insert(
     0, os.path.join(os.path.dirname(os.path.abspath(__file__)), "src"))
 
 # Import third-party modules
-from rezbuild.builder import PythonSourceBuilder
+from rezbuild import PythonSourceBuilder
 
 
 if __name__ == '__main__':